move width and height into mangostore
